[bt] Stop storing unparsed advertising data.
Parse advertising data and EIR data as soon as we receive it.
Add flags to the parsed advertising data.
Refactor DiscoveryFilter to use the parsed data, which will be null if
the data failed to parse.

Add services that are in the EIR for BR/EDR peers to the peer's
services.

Bug: b/294861923
Bug: fxbug.dev/85368
Test: fx test bt-host-gap-tests

Change-Id: Iaf45def855c8b3ccdd8ff5082d2fb5e802c16e58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895781
Reviewed-by: Lulu Wang <luluwang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>